More documentation, provide dataset download script
Add documentation for how to use a NeMo Docker container as a basis for the installation. Though, this is not compatible for the ctc_decoder installation of
the NeMo ctc_decoder installation script which requires Python 3.7 and there is
no current NeMo container with such an old Python version.

Add dataset download script for downloading audio datasets, e.g. the raddromur
ASR audio dataset.